+++ src/widgets/Checkbox.cpp
@@ -67,12 +67,12 @@ namespace pTK
         Size size = getSize();
         Point pos = getPosition();
         
-        float posy = pos.y + size.height/2 + size.height*0.25f;
-        float posx = pos.x + size.width/2 - size.height*0.20f;
-        float longy = size.height*0.7f*0.9f;
-        float shortx = size.height*0.3f*0.9f;
+        float posy = (float)pos.y + (float)size.height/2.0f + (float)size.height*0.25f;
+        float posx = (float)pos.x + (float)size.width/2.0f - (float)size.height*0.20f;
+        float longy = (float)size.height*0.7f*0.9f;
+        float shortx = (float)size.height*0.3f*0.9f;
         float cornerRadius = getCornerRadius();
-        float outline = getOutlineThickness()*2;
+        float outline = getOutlineThickness()*2.0f;
         
         // Setup clip rects.
         SkRect rect;
